requirements.txt
kafka-python==2.0.2
faker==24.0.0
pyspark==3.5.0
delta-spark==3.1.0
deltalake==0.17.4
streamlit==1.32.0
plotly==5.20.0
pandas==2.2.1
